Remove Semaphore and overloads that use it.
Semaphore might get added back in a future version, but for now it makes
little sense. The Vulkan backend only uses a single VkQueue, so
submitted commands are already implicitly ordered based on that. In the
future, this might change (and multiple queues used), in which case
Semaphore could be useful. However, there's some questions about how it
could be implemented for Metal -- it probably couldn't.