Allow uniform buffers to be bound with an offset+size.
* There is a new BindableResource subclass called DeviceBufferRange, which is a thin wrapper over a DeviceBuffer, adding an offset and a size. This can be used to bind different regions of the same buffer to different ResourceSets (or the same set).

Forthcoming (with some groundwork now in place) is the ability to bind ResourceSets with a dynamic set of buffer offsets determined when binding to a CommandList.

To go along with this, there are two new GraphicsDevice properties:

* UniformBufferMinOffsetAlignment
* StructuredBufferMinOffsetAlignment

These values denote the required byte alignment for uniform and structured buffer offsets. When using a DeviceBufferRange in a ResourceSetDescription, its offset MUST be a multiple of one of these properties, depending on the type of resource (uniform or structured buffer).

It is possible for this feature to be unavailable (potentially on D3D11 on some Windows 7 drivers). GraphicsDeviceFeatures.BufferRangeBinding identifies this capability.