Implement BindableResource on Texture directly.
A Texture can now be included directly in a ResourceSet. If you do this, it behaves identically to a TextureView which covers the Texture's full mip and array layer range, with the same PixelFormat. This is largely a convenience feature, as most TextureViews cover the full range of a Texture in practice.