Add a Veldrid.RenderDoc, a helper library for controlling RenderDoc.
RenderDoc exposes a simple in-application API that can be used to
configure and trigger captures, as well as launch and manage the replay
UI application.

To use this new functionality, call RenderDoc.Load(), optionally with a
path to the RenderDoc shared library. Then, you can call TriggerCapture
and LaunchReplayUI, along with a variety of other functions.