Add reference counting to Vulkan resources.
This improves the usability of Vulkan by allowing resources to be
Dispose'd while they are still in use by a submitted CommandList.
An in-progress CommandList will retain a reference to each resource
it uses, and the references will be released once the work item
completes. This is how other backends work implicitly.