[Vulkan] Fix an issue with image layout transitions.
If an image was written to in a compute shader and then subsequently
read as a read-only texture, it could be used in an invalid layout in
some cases. This change adds additional tracking so that this scenario
is avoided, but involves more defensive transitions.

Added a test to validate this behavior.