[Vulkan] Round allocation sizes to the nearest multiple of bufferImageGranularity.
This is a conservative fix for the "memory aliasing" validation warning that can pop up in some situations.

This approach will add some allocation overhead for many small allocations, but should be very minimal in real applications. On a machine with a bufferImageGranularity of 1024, NeoDemo uses less than 1 MB of additional memory with this change.